+++ api/urls.py
@@ -6,8 +6,8 @@ from api.views.students import (
     StudentDetailView,
     StudentCreateView
 )
-from api.views.courses import SectionListView, SectionDetailView
-
+from api.views.sections import SectionListView, SectionDetailView
+from api.views.courses import CourseListAPIView, CourseDetailAPIView
 
 urlpatterns = [
     path('teachers/', TeacherAPIView.as_view(), name="teacher_list"),
@@ -15,6 +15,8 @@ urlpatterns = [
     path('students/', StudentAPIView.as_view(), name="student_list"),
     path('students/<code>', StudentDetailView.as_view()),
     path('student/create', StudentCreateView.as_view(), name="student_create"),
+    path('courses/create/', CourseListAPIView.as_view(), name="course_create"),
+    path('courses/<id>', CourseDetailAPIView.as_view(), name="course_detail"),
     path('courses/sections/', SectionListView.as_view(), name="section_list"),
     path('courses/sections/<int:pk>/', SectionDetailView.as_view(), name="section_detail"),
 ]
